Correct Data type conversion.
git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk/edk2@10981 6f19259b-4bc3-4df7-8a09-765794883524

+++ MdeModulePkg/Universal/SetupBrowserDxe/Ui.c
@@ -1134,7 +1134,7 @@ GetWidth (
     Width -= SUBTITLE_INDENT;
   }
 
-  return Width - LEFT_SKIPPED_COLUMNS;
+  return (UINT16) (Width - LEFT_SKIPPED_COLUMNS);
 }
 
 /**
